+++ doc/changelog.md
@@ -10,6 +10,7 @@ arguments will be bound to (e.g., main(args){}); any arguments not processed by 
 the identifier specified in main contained as an array of strings, if provided. Command line arguments prefixed with a dash or double-dash
 shall be prefixed with a backward slash to be contained in the array of strings bound by the specified identifier - otherwise they will be ignored.
 - Added ID primitive function. A call to ID returns the memory address of its input as a string.
+- Integrated a FileContainer and respective functions - open, close, and readLine.
 ### Changes:
 - The Preprocessor now handles include directives using a stack. It is similar to the C Preprocessor
 in which it will search in the following manner: "In the directories of the currently opened include files, in the reverse order in which they were opened. The search begins in the directory of the